ONLY AUTH VIEW + LOGOUT+Rearrangement in compontents
